Adding Multicast control package
This is an implementation of LoRaWAN Remote Multicast Setup
Specification v1.0.0.
It is supposed to be used as an on-demand plug-in/application extension.
Basic operation of the control package includes parsing of multicast
control commands (traffic on port 200), their interpretation and
generation of appropriate responses. Before passing in any control
commands, the application must activate the package by feeding it the
root key which will be used to derive multicast key encryption key (McKEKey).
This McKEKey is then ued to derive network and application session keys
when the multicast address is conveyed to the device using
MCAST_GROUP_SETUP request. The root key in case of LoRaWAN 1.0.X is a
special out-of-band pre-provisioned key. For v1.1.X AppKey is used as
root key for this plugin too.
The application is responsible for the transmission of responses at port
200 which are generated by the plug-in.